more consistent cache
